fixed authOptions and 'Invalid Compact JWE' error
