+++ src/app/sender/page.jsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../lib/auth";
 
 export default async function Home() {
   const session = await getServerSession(authOptions);
